refactor/auth-pages: extrai subcomponentes e dados de apresentação
- PasswordStrengthIndicator: extrai função de cálculo, config e JSX do CadastroForm
- RecuperarSenhaSuccess: extrai tela de sucesso em componente separado
- _data/features.ts: move array de features do AuthLayout para módulo dedicado
